Button text color change support.
Hello, 

I am trying to change my button text color but I don't find any property for this... so I have added a few lines of code in "Button.js"  for button text color support. Please suggest this is correct? it's working for me.  

If any existing way in this UI kit so please suggest.